Moved SP propeller models to SP folder
