--- gpkitmodels/GP/aircraft/prop/dae51polars/dae51.txt
@@ -1,81 +1,0 @@
-1.000000 -0.000051
-   0.989166  0.001996
-   0.975819  0.004501
-   0.957448  0.007991
-   0.931599  0.013004
-   0.901654  0.018957
-   0.870818  0.025167
-   0.839998  0.031363
-   0.809336  0.037430
-   0.778839  0.043301
-   0.748503  0.048933
-   0.718326  0.054243
-   0.688314  0.059175
-   0.658443  0.063716
-   0.628653  0.067839
-   0.598855  0.071530
-   0.569004  0.074792
-   0.539074  0.077652
-   0.509071  0.080095
-   0.479015  0.082100
-   0.448934  0.083641
-   0.418877  0.084690
-   0.388863  0.085217
-   0.358908  0.085197
-   0.329049  0.084600
-   0.299321  0.083399
-   0.269760  0.081570
-   0.240419  0.079081
-   0.211388  0.075895
-   0.182766  0.071967
-   0.154699  0.067243
-   0.127404  0.061670
-   0.101193  0.055204
-   0.076503  0.047832
-   0.054060  0.039675
-   0.034881  0.031085
-   0.020059  0.022737
-   0.010052  0.015405
-   0.004208  0.009557
-   0.001251  0.005050
-   0.000113  0.001479
-   0.000123 -0.001494
-   0.001002 -0.004085
-   0.002793 -0.006422
-   0.006019 -0.008524
-   0.011960 -0.010609
-   0.022698 -0.012727
-   0.040333 -0.014554
-   0.064233 -0.015717
-   0.091882 -0.016173
-   0.121305 -0.016067
-   0.151560 -0.015570
-   0.182231 -0.014790
-   0.213149 -0.013797
-   0.244240 -0.012649
-   0.275464 -0.011392
-   0.306798 -0.010061
-   0.338230 -0.008690
-   0.369725 -0.007318
-   0.401184 -0.005962
-   0.432597 -0.004640
-   0.463976 -0.003366
-   0.495325 -0.002147
-   0.526644 -0.000994
-   0.557930  0.000082
-   0.589182  0.001071
-   0.620390  0.001960
-   0.651547  0.002731
-   0.682648  0.003367
-   0.713694  0.003847
-   0.744690  0.004152
-   0.775644  0.004269
-   0.806546  0.004190
-   0.837373  0.003898
-   0.868134  0.003359
-   0.898985  0.002532
-   0.929868  0.001410
-   0.956572  0.000219
-   0.975393 -0.000742
-   0.988983 -0.001472
-   1.000000 -0.002051
